added procedure admin

+++ archivosOther/resources/views/Users/index.blade.php
@@ -3,16 +3,18 @@
 @section('content')
 
     <div class="row" style="margin-left: 5%; margin-right: 5%">
-        <h2>Read Data</h2>
+        <h2>Listado de Usuarios</h2>
     </div>
+    @can('users.create')
     <div class="row" style="margin-left: 5%; margin-right: 5%">
-        <a class="btn btn-primary" href="users/create" style="margin-bottom: 15px;">Create New</a>
+        <a class="btn btn-primary" href="users/create" style="margin-bottom: 15px;">Crear Nuevo Usuario</a>
       @if(Session::has('message'))
       <div class="alert-custom">
           <p>{!! Session('message') !!}</p>
       </div>
       @endif()
     </div>
+    @endcan
 
     <div class="row" style="margin-left: 5%; margin-right: 5%">
 
@@ -22,9 +24,9 @@
                 <th style="padding-left: 15px;">#</th>
                 <th>Name</th>
                 <th>Email</th>
-                <th>Facebook</th>
-                <th>Mobile</th>
-                <th width="110px;">Action</th>
+                <th>Direccion</th>
+                <th>Telefono</th>
+                <th width="110px;">Accion</th>
             </tr>
             </thead>
             <tbody>
@@ -34,18 +36,21 @@
                     <td style="padding-left: 15px;">{!! $user->id !!}</td>
                     <td>{!! $user->name !!}</td>
                     <td>{!! $user->email !!}</td>
-                    <td>{!! $user->fb !!}</td>
-                    <td>{!! $user->mobile !!}</td>
+                    <td>{!! $user->address !!}</td>
+                    <td>{!! $user->phone !!}</td>
                     <td>
+                    @can('users.edit')
                         <a class="btn btn-success btn-sm" href="users/{!! $user->id !!}/edit" title="Editar">
-                            <i class="zmdi zmdi-edit"></i>&nbsp;Edit</a>
+                            <i class="zmdi zmdi-edit"></i></a>
+                    @endcan
+                    @can('users.destroy')
+                        <a class="btn btn-danger btn-sm" href="/users/{!! $user->id !!}" onclick="event.preventDefault();
+                            document.getElementById('deleteForm').submit();" title="Eliminar"><i class="zmdi zmdi-delete"></i></a>
 
-                        <a class="btn btn-danger btn-sm" href="/users/{!!$user->id_new!!}" onclick="event.preventDefault();
-                            document.getElementById('deleteForm').submit();" title="Eliminar"><i class="zmdi zmdi-delete"></i>&nbsp;Delete</a>
-
-                        {!! Form::open(['id' => 'deleteForm', 'method' => 'DELETE', 'url' => '/users/' . $user->id]) !!}
+                        {!! Form::open(['id' => 'deleteForm', 'method' => 'DELETE', 'url' => '/users/' . $user->id ]) !!}
                         {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm', 'style'=> 'display:none']) !!}
                         {!! Form::close() !!}
+                    @endcan
                     </td>
                 </tr>
             @endforeach
